#14 added support for AAVE assets on Polygon, Avalanche and Optimism

+++ src/main/java/ch/grignola/service/scanner/aave/model/User.java
@@ -3,10 +3,12 @@ package ch.grignola.service.scanner.aave.model;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.quarkus.runtime.annotations.RegisterForReflection;
 
+import java.math.BigDecimal;
+
 @RegisterForReflection
 public class User {
 
     @JsonProperty("unclaimedRewards")
-    public long unclaimedRewards;
+    public BigDecimal unclaimedRewards;
 
 }
